Restructure tests directory (#117)
* Init

* More refactoring

* rename dbTestEnv

* Clean up package.json

* Fix file path, add 'he' types package

* Add tests-specific README, some more restructuring
