+++ package.json
@@ -27,9 +27,7 @@
     "tsc": "tsc",
     "-testing-": "--------testing--------",
     "test": "npx jest --verbose --testPathIgnorePatterns='(seq|reg|git).[jt]s'",
-    "regtest": "jest --projects tests/reg_tests",
-    "dbtest": "jest -i --projects tests/ --verbose",
-    "prodtest": "PROD=true jest reg.[jt]s",
+    "dbtest": "jest -i --projects tests/database --verbose",
     "coverage": "./node_modules/jest-cli/bin/jest.js --coverage",
     "-building-": "--------building--------",
     "build_backend": "rm -rf dist; mkdir -p dist; babel --extensions '.js,.ts' . -d dist/ --copy-files --ignore node_modules --include-dotfiles; rm -rf dist/.git",
@@ -145,6 +143,7 @@
     "@types/cookie": "^0.4.1",
     "@types/domutils": "^1.7.4",
     "@types/fs-extra": "^9.0.13",
+    "@types/he": "^1.1.2",
     "@types/htmlparser2": "^3.10.3",
     "@types/jest": "^27.0.0",
     "@types/jsonwebtoken": "^8.5.6",
