+++ tests/database/search.test.seq.ts
@@ -1,6 +1,6 @@
-import searcher from "../services/searcher";
-import prisma from "../services/prisma";
-import { LeafQuery, ParsedQuery } from "../types/searchTypes";
+import searcher from "../../services/searcher";
+import prisma from "../../services/prisma";
+import { LeafQuery, ParsedQuery } from "../../types/searchTypes";
 
 beforeAll(async () => {
   searcher.subjects = {};
