+++ tests/database/updater.test.seq.ts
@@ -1,16 +1,16 @@
 import _ from "lodash";
-import Updater from "../services/updater";
+import Updater from "../../services/updater";
 import {
   Course as CourseType,
   Section as SectionType,
   Requisite,
-} from "../types/types";
-import prisma from "../services/prisma";
-import { Prisma } from "prisma";
-import Keys from "../utils/keys";
-import dumpProcessor from "../services/dumpProcessor";
-import termParser from "../scrapers/classes/parsersxe/termParser";
-import elasticInstance from "../utils/elastic";
+} from "../../types/types";
+import prisma from "../../services/prisma";
+import { Prisma } from "@prisma/client";
+import Keys from "../../utils/keys";
+import dumpProcessor from "../../services/dumpProcessor";
+import termParser from "../../scrapers/classes/parsersxe/termParser";
+import elasticInstance from "../../utils/elastic";
 
 function processCourse(classInfo: any): Prisma.CourseCreateInput {
   const additionalProps = {
