+++ tests/end_to_end/jest.config.js
@@ -1,13 +1,8 @@
 module.exports = {
-  name: "workflow",
+  name: "End to end",
   displayName: "Tests for GitHub Workflows",
   rootDir: "./",
   testEnvironment: "node",
   moduleFileExtensions: ["js", "json", "node", "ts"],
   testMatch: ["**/*.(spec|test).git.[jt]s?(x)"],
-  testPathIgnorePatterns: [
-    "<rootDir>/tests/workflow_tests/jest.config.js",
-    "<rootDir>/node_modules",
-    "<rootDir>/dist/",
-  ],
 };
