--- tests/reg_tests/README.md
@@ -1,1 +1,0 @@
-These test are all outdated.
