--- tests/reg_tests/jestSetupFile.js
@@ -1,1 +1,0 @@
-import "regenerator-runtime/runtime";
